fixes getting rid of useles or wrong prints
for some rease there were fprintf statement i thought that was weird so i got rid of them as well as all useles print statement fixed sme so that their all newline got rid of the thing that always takes you to the previous cluster ecause apperetly thats not supposed to be there or something i dont know in the announcment their was something about how this didnt need to be there i was confused by it. Thats the main reason im writing this anyway. put in some breaks to while loop that wernt put in